TMI Mat extension
big changes to config.jl
