+++ Project.toml
@@ -30,10 +30,12 @@ UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 COSMO = "1e616198-aa4e-51ec-90a2-23f7fbd31d8d"
 GeoPythonPlot = "1e05a8e1-7dec-4f9e-9d3d-7df52321841b"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+MAT = "23992714-dd62-5051-b70f-ba57cb901cac"
 
 [extensions]
 TMIGeoPythonPlotExt = "GeoPythonPlot"
 TMIJuMPExt = ["JuMP", "COSMO"]
+TMI_MAT_Ext = "MAT"
 
 [compat]
 COSMO = "0.8"
